requirements.txt
torch>=1.13.0
transformers>=4.30.0
torchvision>=0.14.0
numpy
pandas
Pillow
matplotlib
nltk
tqdm
scikit-learn
